+++ source/core/graphics/pipeline/Pipeline.cpp
@@ -63,14 +63,18 @@ void CPipelineBase::reCreate(vk::RenderPass& renderPass, uint32_t _subpass)
 
 void CPipelineBase::cleanup()
 {
-    if(m_descriptorSetLayout)
-        CDevice::inst()->destroy(m_descriptorSetLayout);
-    if(m_descriptorPool)
-        CDevice::inst()->destroy(m_descriptorPool);
-    if(m_pipeline)
-        CDevice::inst()->destroy(m_pipeline);
-    if(m_pipelineLayout)
-        CDevice::inst()->destroy(m_pipelineLayout);
+    if(!bIsClean)
+    {
+        if(m_descriptorSetLayout)
+            CDevice::inst()->destroy(m_descriptorSetLayout);
+        if(m_descriptorPool)
+            CDevice::inst()->destroy(m_descriptorPool);
+        if(m_pipeline)
+            CDevice::inst()->destroy(m_pipeline);
+        if(m_pipelineLayout)
+            CDevice::inst()->destroy(m_pipelineLayout);
+        bIsClean = true;
+    }
 }
 
 void CPipelineBase::bind(vk::CommandBuffer &commandBuffer)
